requirements.txt
openai
moviepy==1.0.3
pillow
feedparser
python-dotenv
requests
google-api-python-client